+++ src/ui/editor/areas/editors.ts
@@ -4,6 +4,7 @@ import IslandEditor from './island/IslandEditor';
 import LayoutsEditor from './layouts/LayoutsEditor';
 import IsoGridEditor from './isoGrid/IsoGridEditor';
 import { getParams } from '../../../params';
+import InventoryEditor from './model/InventoryEditor';
 
 const isLBA1 = getParams().game === 'lba1';
 
@@ -11,12 +12,14 @@ const editors = isLBA1
 ? [
     GameplayEditor,
     ModelEditor,
+    InventoryEditor,
     IsoGridEditor,
     LayoutsEditor
 ]
 : [
     GameplayEditor,
     ModelEditor,
+    InventoryEditor,
     IslandEditor,
     IsoGridEditor,
     LayoutsEditor